Use StateHandler and Controller
